+++ internal/app/osd/proto/api.proto
@@ -9,7 +9,6 @@ import "google/protobuf/empty.proto";
 //
 // OSD Service also implements all the API of Init Service
 service OSD {
-  rpc DF(google.protobuf.Empty) returns (DFReply) {}
   rpc Dmesg(google.protobuf.Empty) returns (Data) {}
   rpc Kubeconfig(google.protobuf.Empty) returns (Data) {}
   rpc Logs(LogsRequest) returns (stream Data) {}
@@ -156,14 +155,3 @@ message TopRequest {}
 message TopReply { ProcessList process_list = 1; }
 
 message ProcessList { bytes bytes = 1; }
-
-// The response message containing the requested df stats.
-message DFReply { repeated DFStat stats = 1; }
-
-// The response message containing the requested processes.
-message DFStat {
-  string filesystem = 1;
-  uint64 size = 2;
-  uint64 available = 3;
-  string mounted_on = 4;
-}
